Add draft of database data structure
Not sure the name database even makes sense here. This is just the data
model I want for some personal data storage / entry. The plan is to
possibly migrate this to sqlite at some point, but for now I don't plan
to use this for large amounts of data, so this won't need to be optimized
a ton. I'm also planning to just store it in a JSON file and possibly
overwrite the whole file whenever something changes.